backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
yfinance
pandas
numpy
google-generativeai
openai
python-dotenv
requests
beautifulsoup4
sqlalchemy
asyncpg
python-multipart
python-jose[cryptography]
passlib[bcrypt]
ta
pandas_ta